fix: move shared function components, into shared component dir
