+++ src/shared/components/functions/FunctionsList/GroupedFunctionsList.tsx
@@ -5,7 +5,7 @@ import {FLUX_FUNCTIONS} from 'src/shared/constants/fluxFunctions'
 import {FluxToolbarFunction} from 'src/types/shared'
 import Fn from 'src/flows/shared/FilterList/InjectionOption'
 import SearchWidget from 'src/shared/components/search_widget/SearchWidget'
-import FunctionTooltipContent from 'src/flows/pipes/RawFluxEditor/FunctionsList/perFunction/FunctionToolTipContent'
+import FunctionTooltipContent from 'src/shared/components/functions/perFunction/FunctionToolTipContent'
 
 interface Props {
   onSelect: (fn: FluxToolbarFunction) => void
